+++ ipyexperiments/utils/mem.py
@@ -1,51 +1,62 @@
 "Helper utility functions for memory management"
 
-import threading
-import time
+import threading, time, gc
 from collections import namedtuple
 from .inject_pynvx import get_pynvml
 
 pynvml = None
-have_cuda = 0
+use_gpu = 0
 
 try:
     # currently relying on pytorch
     import torch
+    use_gpu = torch.cuda.is_available()
 
     pynvml = get_pynvml()
     pynvml.nvmlInit()
-    have_cuda = 1
 except:
-    raise RuntimeError(
-        "these functions require NVIDIA environment, check the output of `nvidia-smi`") from None
+    raise RuntimeError("these functions require NVIDIA environment, check the output of `nvidia-smi`") from None
 
 ############# gpu memory helper functions ############
 
 GPUMemory = namedtuple('GPUMemory', ['total', 'free', 'used'])
 
 def preload_pytorch():
+    """ Do a small operation on CUDA to get the pytorch/cuda structures in place.
+
+    A must to be run first if you're going to compare any CUDA-related numbers.
+    """
     torch.ones((1, 1)).cuda()
 
 preload_pytorch()  # needed to run first to get the measurements right
-
-def gpu_mem_allocate_mbs(n):
-    " allocate n MBs, return the var holding it on success, None on failure "
-    try:
-        d = int(2**9*n**0.5)
-        return torch.ones((d, d)).cuda().contiguous()
-    except:
-        return None
 
 def b2mb(num):
     """ convert Bs to MBs and round down """
     return int(num/2**20)
 
+def gpu_mem_allocate_mbs(n, fatal=False):
+    """ Try to allocate n MBs.
+
+    Return the variable  holding it on success, None on failure.
+
+    Ask for at least 6MB or more.
+
+    fatal=True will throw an exception on failure to allocate (default is False).
+    """
+    # don't try to allocate less than 6MB as it'd be imprecise, need to probably switch to bytes allocation
+    if n < 6: return None # XXX: FIXME
+    try:
+        d = int(2**9*n**0.5)
+        return torch.ones((d, d)).cuda().contiguous()
+    except Exception as e:
+        if not fatal: return None
+        raise e
+
 # for gpu returns GPUMemory(total, free, used)
-# for cpu returns GPUMemory(0, 0, 0)
 # for invalid gpu id returns GPUMemory(0, 0, 0)
 def gpu_mem_get_mbs(id=None):
-    "query nvidia for total, used and free memory for gpu in MBs. if id is not passed, currently selected torch device is used"
-    if not have_cuda:
+    """ Query nvidia for total, used and free memory for gpu in MBs. if id is not passed, currently selected torch device is used """
+    if not use_gpu:
         return GPUMemory(0, 0, 0)
     if id is None:
         id = torch.cuda.current_device()
@@ -56,19 +67,39 @@ def gpu_mem_get_mbs(id=None):
     except:
         return GPUMemory(0, 0, 0)
 
-def gpu_mem_get_free_mbs():  return gpu_mem_get_mbs().free
-def gpu_mem_get_used_mbs():  return gpu_mem_get_mbs().used
-def gpu_mem_get_total_mbs(): return gpu_mem_get_mbs().total
+def gpu_mem_get_total_mbs():
+    """ Return the amount of total memory (in rounded MBs) """
+    return gpu_mem_get_mbs().total
+
+def gpu_mem_get_free_mbs():
+    """ Return the amount of free memory (in rounded MBs) """
+    return gpu_mem_get_mbs().free
+
+def gpu_mem_get_free_no_cache():
+    """ Return the amount of free memory after flushing caching (in rounded MBs) """
+    gc.collect()
+    torch.cuda.empty_cache()
+    return gpu_mem_get_free_mbs()
+
+def gpu_mem_get_used_mbs():
+    """ Return the amount of used memory (in rounded MBs) """
+    return gpu_mem_get_mbs().used
+
+def gpu_mem_get_used_no_cache():
+    """ Return the amount of used memory after flushing caching (in rounded MBs) """
+    gc.collect()
+    torch.cuda.empty_cache()
+    return gpu_mem_get_used_mbs()
 
 def gpu_mem_leave_free_mbs(n):
     """Consume whatever memory is needed so that n MBs are left free.
 
     On success it returns a variable that holds the allocated memory, which
     needs to be kept alive as long as it's needed to hold that memory. Call
-    `del` to release the memory when that memory is no longer needed.
+    `del` to release the memory when it is no longer needed.
 
-    This is very useful if the test needs to hit OOM, so this function will
-    leave just the requested amount of GPU RAM free, regardless of GPU
+    This function is very useful if the test needs to hit OOM, so this function
+    will leave just the requested amount of GPU RAM free, regardless of GPU
     utilization or size of the card.
 
     """